+++ gradle-plugin/src/main/kotlin/dev/zacsweers/metro/gradle/MetroPluginExtension.kt
@@ -446,6 +446,21 @@ constructor(
     objects.booleanProperty("metro.enableFunctionProviders", false)
 
   /**
+   * Enable/disable [kotlin.reflect.KClass]/[Class] interop for multibinding map keys. When enabled,
+   * `java.lang.Class` and `kotlin.reflect.KClass` are treated as interchangeable in map key types,
+   * matching Kotlin's own annotation compilation behavior. This only applies to map keys because
+   * these are the only scenario where annotation arguments are materialized into non-annotation
+   * code (i.e. `@ClassKey(Foo::class) -> Map<Class<*>, V>`).
+   *
+   * Disabled by default because this is purely for annotations interop and potentially comes at
+   * some runtime overhead cost to interop since `KClass` types are still used under the hood and
+   * must be mapped in some cases. It's recommended to migrate these to `KClass` and call `.java`
+   * where necessary if possible.
+   */
+  public val enableKClassToClassMapKeyInterop: Property<Boolean> =
+    objects.booleanProperty("metro.enableKClassToClassMapKeyInterop", false)
+
+  /**
    * If set, the Metro compiler will dump verbose report diagnostics about resolved dependency
    * graphs to the given destination. Outputs are per-compilation granularity (i.e.
    * `build/metro/main/...`).