+++ CHANGELOG.md
@@ -6,6 +6,12 @@ Changelog
 
 ### New
 
+- **`Class`/`KClass` map key interop**
+
+This release introduces a special-cased opt-in `java.lang.Class` and `kotlin.reflect.KClass` interop on JVM/android compilations. While these types are not intrinsics of each other in regular code, they _are_ in annotations and are often used in `Map` multibindings. Metro can support these if you enable the `enableKClassToClassMapKeyInterop` option. When enabled, `java.lang.Class` and `kotlin.reflect.KClass` are treated as interchangeable in map key types, matching Kotlin's own annotation compilation behavior. This only applies to map keys because these are the only scenario where annotation arguments are materialized into non-annotation code (i.e. `@ClassKey(Foo::class) -> Map<Class<*>, V>`).
+
+This is disabled by default (even if other framework interops like `includeDagger` are enabled) because this is purely for annotations interop and potentially comes at some runtime overhead cost to interop since `KClass` types are still used under the hood and must be mapped in some cases. It's recommended to migrate these to `KClass` and call `.java` where necessary if possible.
+
 ### Enhancements
 
 - **[FIR]**: Report adhoc graph extension factories as these are unsupported in Metro (but apparently supported in Dagger!)